requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-admin-tools==0.8.0
django-adminactions==1.4
django==1.9.12
future==0.16.0            # via xlwt-future
pytz==2016.10             # via django-adminactions
six==1.10.0               # via django-adminactions
unicodecsv==0.14.1        # via django-adminactions
xlrd==1.0.0               # via django-adminactions
xlwt-future==0.8.0        # via django-adminactions
